Fix access_token authentication
Greenhouse API uses user/password authentication even if the  service
is authenticating with an API token. We were using token auth, which
caused the headers to be invalid. This commit changes the api_token
option to authenticate as a user/password with an empty password. It
also changes the default endpoint to the url for the harvest API, which
is the one we use.